Add optional ASSERTION_URL setting
We pass the result of `get_current_domain` to pysaml2 to validate
the source of the incoming StatusResponse (the scheme+host of the
request).

This is not viable in a reverse proxy scenario, in particular when the
host, port, and/or scheme of the django server changes with churn.

As a solution, alter `get_current_domain` to just use a setting
variable, `ASSERTION_URL`, instead of the scheme and host of the
incoming SAML request. If the setting is not present, use the
current behavior.